style : 매서드 이동
